chore: full health scan and VS Code setup — PyForge v0.2.0
Critical fix:
- Fix broken benchmarks after pyforge-core refactor (Integer i32→i64)

VS Code workspace:
- settings.json: rust-analyzer + ruff + format-on-save
- extensions.json: 7 recommended extensions
- tasks.json: cargo check/clippy/test + full check
- launch.json: Rust debugger (CodeLLDB) + Python debugger

Reports:
- HEALTH_REPORT.md: 11 issues found (2 critical/high fixed, 9 documented)
- INNOVATION_PROPOSALS.md: 10 proposals ranked by impact
  Top 3: queryset-to-Rust batch path (game-changing),
  pyforge doctor command (high), auto-schema from dataclass (high)

+++ pyforge-django/benches/bench_field_types.rs
@@ -51,7 +51,7 @@ fn bench_integerfield(c: &mut Criterion) {
     let (d, v) = make_single(
         "score",
         DjangoFieldType::IntegerField,
-        FieldValue::Integer(750),
+        FieldValue::Integer(750i64),
     );
     c.bench_function("field_type/IntegerField", |b| {
         b.iter(|| serialize_model_fields(black_box(&d), black_box(&v)))
